Move extension appropriately
