+++ ios-app/New Group/UIDevice+ModelName.swift
@@ -1,0 +1,21 @@
+//
+//  UIDevice+ModelName.swift
+//  ios-app
+//
+//  Created by Karthik raja on 9/8/19.
+//  Copyright © 2019 Testpress. All rights reserved.
+//
+import UIKit
+
+extension UIDevice {
+    var modelName: String {
+        var systemInfo = utsname()
+        uname(&systemInfo)
+        let machineMirror = Mirror(reflecting: systemInfo.machine)
+        let identifier = machineMirror.children.reduce("") { identifier, element in
+            guard let value = element.value as? Int8, value != 0 else { return identifier }
+            return identifier + String(UnicodeScalar(UInt8(value)))
+        }
+        return identifier
+    }
+}
